build(69cdcf11c083b16d6deb4dc9748c8912c1339981): Update gamejam.html.md
# Blah blah blah
#
# ### Unsorted list:
#
# * A link: [link title](http://example.com/url)
# * An inline image: ![alt text](https://example.com/image.png "Image Title")
#
# ### Sorted list:
#
# 1. An `inline code`.
# 2. Be **bold** not *italic*.
#
# ### Code block:
#
# ```python
# import this
# ```
#
# See more at [Markdown Cheatsheet](https://github.com/adam-p/markdown-here/wiki/Markdown-Cheatsheet).
CSUEB GAME JAMS are a series of three events set to occur in the middle of each quarter of the 2016/17 academic year.  Following a series of weekly tutorials by senior students and industry professionals each game jam is a separate culminating event which takes place over a weekend. Each GAME JAM opens on friday afternoon with a mixer event, a keynote which outlines and unveils the common theme for the particular event, a idea pitching session, that is followed by team formation.  On Saturday and Sunday, teams arrive in the morning for a small breakfast and peptalk and begin to work on their ideas.  Throughout the day, industry and professional participants circulate among the teams and offer mentoring and advice. On Sunday at 3:00PM it is pencils down and teams prepare to present their projects to the public.  The event concludes with presentations of the projects and a celebration of accomplishments.  The GAME JAM is a non-competitive event focussed on ideation and innovation and creating prototypes, rather than on product oriented success.